two-phase flow specific BC moved to proteus.mprans
